app/requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
virtualenv==13.1.2
Werkzeug==0.11.3
wheel==0.26.0
flask-sqlalchemy==2.1
flask-script==2.0.5
psycopg2==2.6.1
ipython==4.0.3
flask-migrate==1.7.0
fake-factory==0.5.3
factory-boy==2.6.1
marshmallow==2.6.0
